Folder with ejercicio00 created
