+++ Ejercicio00/.gitignore
