+++ Ejercicio00/src/main/java/com/mycompany/entrega1/exceptions/LuminosidadExcedidaException.java
